requirements.txt
pyrogram
tgcrypto
yt-dlp
pymongo
python-dotenv